Move `shared/components/FieldRadio` to `design/FieldRadio`
